Move help messages to separate source/header
