+++ GPTifier/include/help_messages.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+void print_help_messages();
